refactor: 콘솔 패키지 구분
