+++ src/main/java/org/prgms/voucherProgram/SpringVoucherApplication.java
@@ -1,6 +1,6 @@
 package org.prgms.voucherProgram;
 
-import org.prgms.voucherProgram.domain.ConsoleProgram;
+import org.prgms.voucherProgram.console.ConsoleProgram;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
