+++ src/main/java/org/prgms/voucherProgram/console/menu/VoucherMenuType.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.domain.menu;
+package org.prgms.voucherProgram.console.menu;
 
 import java.util.stream.Stream;
