+++ src/main/java/org/prgms/voucherProgram/console/program/VoucherProgram.java
@@ -1,8 +1,11 @@
-package org.prgms.voucherProgram.domain.program;
+package org.prgms.voucherProgram.console.program;
 
 import java.util.UUID;
 
-import org.prgms.voucherProgram.domain.menu.VoucherMenuType;
+import org.prgms.voucherProgram.console.menu.VoucherMenuType;
+import org.prgms.voucherProgram.console.view.Console;
+import org.prgms.voucherProgram.console.view.InputView;
+import org.prgms.voucherProgram.console.view.OutputView;
 import org.prgms.voucherProgram.domain.voucher.VoucherType;
 import org.prgms.voucherProgram.dto.VoucherRequest;
 import org.prgms.voucherProgram.dto.WalletRequest;
@@ -11,9 +14,6 @@ import org.prgms.voucherProgram.exception.WrongCommandException;
 import org.prgms.voucherProgram.exception.WrongDiscountValueException;
 import org.prgms.voucherProgram.exception.WrongEmailException;
 import org.prgms.voucherProgram.service.VoucherService;
-import org.prgms.voucherProgram.view.Console;
-import org.prgms.voucherProgram.view.InputView;
-import org.prgms.voucherProgram.view.OutputView;
 import org.springframework.stereotype.Component;
 
 @Component
