+++ src/main/java/org/prgms/voucherProgram/console/view/Console.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.view;
+package org.prgms.voucherProgram.console.view;
 
 import java.util.List;
 import java.util.Scanner;
